+++ 3rdparty/SUMlib/src/MCMC.cpp
@@ -77,7 +77,7 @@ void MCMC::stepImpl( vector<double>& yNew, double& logLhNew, const size_t i )
 {
    if ( m_krigingUsage == SmartMcmcKriging )
    {
-      calcModel( extendSubSampleToProxyCase( m_pSubSample[i], i ), yNew, m_krigingType ); //calculate expensive yNew
+      calcModel( extendSubSampleToProxyCase( m_pSubSample[i], static_cast<unsigned int>(i) ), yNew, m_krigingType ); //calculate expensive yNew
       logLhNew = calcLh( yNew ); //calculate corresponding log likelihood
    }
    else
